pyproject.toml
[project]
name = "resume-optimizer"
version = "1.0.0"
description = "AI-powered resume optimization using HuggingFace and LlamaIndex"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Resume Optimizer Team"}
]
keywords = ["resume", "ai", "huggingface", "rag", "optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business",
    "Topic :: Multimedia",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "streamlit>=1.28.0",
    "llama-index>=0.9.0",
    "llama-index-llms-huggingface>=0.1.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=4.0.0",
    "pydantic>=2.0.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "huggingface-hub>=0.17.0",
    "accelerate>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"